Help adjusted. No params fixed. Strings added to write option.

+++ dump/formatter/HexdumpFormatter.cpp
@@ -9,7 +9,7 @@ using namespace std;
 using namespace toolslib;
 using namespace toolslib::files;
 
- HexdumpFormatter::HexdumpFormatter(DataFormatter::ByteType type, uint16_t columns, int nAdressSize, CharMode nCharMode)
+ HexdumpFormatter::HexdumpFormatter(DataFormatter::ByteType type, uint16_t columns, uint16_t nAdressSize, CharMode nCharMode)
 : DataFormatter(type, columns, "", "", "", 0)
 , mAddress(0)
 , mAddressSize(nAdressSize)
